requirements.txt
azure-identity
msal
requests
PyQt5
pyjwt
